+++ dev/tests/spec/matchSpec.js
@@ -173,6 +173,7 @@ describe('Match', function(){
 
     });
 
+
     describe('slash is required between word and param', function(){
         
         it('required param after \\w', function(){
@@ -211,4 +212,279 @@ describe('Match', function(){
     
     });
 
+
+    describe('rules', function(){
+        
+        describe('basic rules', function(){
+
+            it('should allow array options', function(){
+
+                var s = crossroads.addRoute('/{foo}/{bar}');
+                
+                s.rules = {
+                    foo : ['lorem-ipsum', 123],
+                    bar : ['dolor', 45]
+                };
+                
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/dolor') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/123') ).toBe( false );
+                expect( s.match('/123/45') ).toBe( true );
+
+            });
+
+            it('should allow RegExp options', function(){
+                
+                var s = crossroads.addRoute('/{foo}/{bar}');
+                
+                s.rules = {
+                    foo : /(^[a-z0-9\-]+$)/,
+                    bar : /(.+)/
+                };
+                
+                
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/dolor') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/45') ).toBe( true );
+            
+            });
+
+            it('should allow function rule', function(){
+                var s = crossroads.addRoute('/{foo}/{bar}/{ipsum}');
+                
+                s.rules = {
+                    foo : function(val, request, params){
+                        return (val === 'lorem-ipsum' || val === 123);
+                    },
+                    bar : function(val, request, params){
+                        return (request !== '/lorem-ipsum');
+                    },
+                    ipsum : function(val, request, params){
+                        return (params.bar === 'dolor' && params.ipsum === 'sit-amet') || (params.bar === 45 && params.ipsum === 67);
+                    }
+                };
+                
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/dolor/sit-amet') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/44/55') ).toBe( false );
+                expect( s.match('/123/45/67') ).toBe( true );
+            });
+
+            it('should work with mixed rules', function(){
+                
+                var s = crossroads.addRoute('/{foo}/{bar}/{ipsum}');
+            
+                s.rules = {
+                    foo : function(val, request, params){
+                        return (val === 'lorem-ipsum' || val === 123);
+                    },
+                    bar : ['dolor', 45],
+                    ipsum : /(sit-amet|67)/
+                };
+                
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/dolor/sit-amet') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/45/67') ).toBe( true );
+
+            });
+
+            it('should only check rules of optional segments if param exists', function(){
+                
+                var a = crossroads.addRoute('/123/:foo:/:bar:');
+                a.rules = {
+                    foo : /^\w+$/,
+                    bar : function(val){
+                        return val === 'ipsum';
+                    }
+                };
+
+                expect( a.match('/123') ).toBe( true );
+                expect( a.match('/123/') ).toBe( true );
+                expect( a.match('/123/asd') ).toBe( true );
+                expect( a.match('/123/asd/') ).toBe( true );
+                expect( a.match('/123/asd/ipsum/') ).toBe( true );
+
+                expect( a.match('/123/asd/45') ).toBe( false );
+                expect( a.match('/123/asd/45/qwe') ).toBe( false );
+                expect( a.match('/123/as#%d&/ipsum') ).toBe( false );
+
+            });
+
+
+            it('should work with shouldTypecast=false', function(){
+                var prevTypecast = crossroads.shouldTypecast;
+                var s = crossroads.addRoute('/{foo}/{bar}/{ipsum}');
+            
+                crossroads.shouldTypecast = false;
+
+                s.rules = {
+                    foo : function(val, request, params){
+                        return (val === 'lorem-ipsum' || val === '123');  //only string validates
+                    },
+                    bar : ['dolor', '45'], //only string validates
+                    ipsum : /(sit-amet|67)/
+                };
+                
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/dolor/sit-amet') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/45/67') ).toBe( true );
+
+                crossroads.shouldTypecast = prevTypecast; //restore
+            });
+
+        });
+
+
+        describe('path alias', function(){
+        
+            it('should work with string pattern', function(){
+            
+                var s = crossroads.addRoute('/{foo}/{bar}/{ipsum}');
+            
+                s.rules = {
+                    0 : function(val, request, params){
+                        return (val === 'lorem-ipsum' || val === 123);
+                    },
+                    1 : function(val, request, params){
+                        return (request !== '/lorem-ipsum');
+                    },
+                    2 : function(val, request, params){
+                        return (params[1] === 'dolor' && params[2] === 'sit-amet') || (params[1] === 45 && params[2] === 67);
+                    }
+                };
+                
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/dolor/sit-amet') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/44/55') ).toBe( false );
+                expect( s.match('/123/45/67') ).toBe( true );
+
+            });
+
+            it('should work with RegExp pattern', function(){
+            
+                var s = crossroads.addRoute(/([\-\w]+)\/([\-\w]+)\/([\-\w]+)/);
+            
+                s.rules = {
+                    0 : function(val, request, params){
+                        return (val === 'lorem-ipsum' || val === 123);
+                    },
+                    1 : function(val, request, params){
+                        return (request !== '/lorem-ipsum');
+                    },
+                    2 : function(val, request, params){
+                        return (params[1] === 'dolor' && params[2] === 'sit-amet') || (params[1] === 45 && params[2] === 67);
+                    }
+                };
+                
+                expect( s.match('/lorem-ipsum') ).toBe( false );
+                expect( s.match('/lorem-ipsum/dolor/sit-amet') ).toBe( true );
+                expect( s.match('lorem-ipsum') ).toBe( false );
+                expect( s.match('/123') ).toBe( false );
+                expect( s.match('123') ).toBe( false );
+                expect( s.match('/123/44/55') ).toBe( false );
+                expect( s.match('/123/45/67') ).toBe( true );
+            
+            });
+
+        });
+
+        describe('special rules', function(){
+        
+            describe('request_', function(){
+                
+                it('should validate whole request', function(){
+                    var s = crossroads.addRoute(/^([a-z0-9]+)$/);
+                    s.rules = {
+                        request_ : function(request){ //this gets executed after all other validations
+                            return request !== 555;
+                        }
+                    };
+                    expect( s.match('lorem') ).toBe( true );
+                    expect( s.match('lorem/dolor/sit-amet') ).toBe( false );
+                    expect( s.match('lorem-ipsum') ).toBe( false );
+                    expect( s.match('123') ).toBe( true );
+                    expect( s.match('555') ).toBe( false );
+                });
+
+                it('should execute after other rules', function(){
+                    var s = crossroads.addRoute('/{foo}/{bar}/{ipsum}');
+                    s.rules = {
+                        foo : function(val, request, params){
+                            return (val === 'lorem-ipsum' || val === 123);
+                        },
+                        bar : ['dolor', 45],
+                        ipsum : /(sit-amet|67|555)/,
+                        request_ : function(request){ //this gets executed after all other validations
+                            return request !== '/123/45/555';
+                        }
+                    };
+                    expect( s.match('/lorem-ipsum') ).toBe( false );
+                    expect( s.match('/lorem-ipsum/dolor/sit-amet') ).toBe( true );
+                    expect( s.match('lorem-ipsum') ).toBe( false );
+                    expect( s.match('/123') ).toBe( false );
+                    expect( s.match('123') ).toBe( false );
+                    expect( s.match('/123/45/67') ).toBe( true );
+                    expect( s.match('/123/45/555') ).toBe( false );
+                });
+            
+                it('can be an array', function(){
+                    var s = crossroads.addRoute(/^([a-z0-9]+)$/);
+                    s.rules = {
+                        request_ : ['lorem', 123]
+                    };
+                    expect( s.match('lorem') ).toBe( true );
+                    expect( s.match('lorem/dolor/sit-amet') ).toBe( false );
+                    expect( s.match('lorem-ipsum') ).toBe( false );
+                    expect( s.match('123') ).toBe( true );
+                    expect( s.match('555') ).toBe( false );
+                });
+
+                it('can be a RegExp', function(){
+                    var s = crossroads.addRoute(/^([a-z0-9]+)$/);
+                    s.rules = {
+                        request_ : /^(lorem|123)$/
+                    };
+                    expect( s.match('lorem') ).toBe( true );
+                    expect( s.match('lorem/dolor/sit-amet') ).toBe( false );
+                    expect( s.match('lorem-ipsum') ).toBe( false );
+                    expect( s.match('123') ).toBe( true );
+                    expect( s.match('555') ).toBe( false );
+                });
+
+                it('should work with optional params', function(){
+                    var s = crossroads.addRoute(':foo:');
+                    s.rules = {
+                        request_ : /^(lorem|123|)$/ //empty also matches!
+                    };
+                    expect( s.match('lorem') ).toBe( true );
+                    expect( s.match('lorem/dolor/sit-amet') ).toBe( false );
+                    expect( s.match('lorem-ipsum') ).toBe( false );
+                    expect( s.match('123') ).toBe( true );
+                    expect( s.match('555') ).toBe( false );
+                    expect( s.match('') ).toBe( true );
+                });
+
+            });
+        
+        });
+
+    });
 });