+++ dev/tests/unit-01.js
@@ -427,10 +427,10 @@ YUI().use('node', 'console', 'test', function (Y){
                 bar : ['dolor', 45]
             };
             
-            Y.Assert.areSame(false, s.match('/lorem-ipsum'));
+            Y.Assert.areSame(true, s.match('/lorem-ipsum'));
             Y.Assert.areSame(true, s.match('/lorem-ipsum/dolor'));
             Y.Assert.areSame(false, s.match('lorem-ipsum'));
-            Y.Assert.areSame(false, s.match('/123'));
+            Y.Assert.areSame(true, s.match('/123'));
             Y.Assert.areSame(false, s.match('123'));
             Y.Assert.areSame(false, s.match('/123/123'));
             Y.Assert.areSame(true, s.match('/123/45'));
@@ -555,10 +555,10 @@ YUI().use('node', 'console', 'test', function (Y){
                 }
             };
             
-            Y.Assert.areSame(false, s.match('/lorem-ipsum'));
+            Y.Assert.areSame(true, s.match('/lorem-ipsum'));
             Y.Assert.areSame(true, s.match('/lorem-ipsum/dolor/sit-amet'));
             Y.Assert.areSame(false, s.match('lorem-ipsum'));
-            Y.Assert.areSame(false, s.match('/123'));
+            Y.Assert.areSame(true, s.match('/123'));
             Y.Assert.areSame(false, s.match('123'));
             Y.Assert.areSame(true, s.match('/123/45/67'));
         },